refactor: utils/view 를 domain 으로  이동
