+++ src/index.js
@@ -1,20 +1,16 @@
 import CarRace from "./domain/CarRace.js";
 import { createCarsFromCarNames } from "./utils/cars.js";
-import {
-  getCarNamesFromUserInput,
-  getTotalRaceCountFromUserInput,
-  printCarRaceWinners,
-} from "./utils/view.js";
+import View from "./domain/View.js";
 
 async function play() {
   // 자동차 이름 입력
-  const carNames = await getCarNamesFromUserInput();
+  const carNames = await View.getCarNamesFromUserInput();
 
   // 자동차 이름으로부터 자동차 생성
   const cars = createCarsFromCarNames(carNames);
 
   // 자동차 경주 횟수 입력
-  const totalRaceCount = await getTotalRaceCountFromUserInput();
+  const totalRaceCount = await View.getTotalRaceCountFromUserInput();
 
   console.log("\n실행 결과");
 
@@ -25,7 +21,7 @@ async function play() {
   carRace.race();
 
   // 우승자 출력
-  printCarRaceWinners(carRace.winners);
+  View.printCarRaceWinners(carRace.winners);
 }
 
 play();
